Fix retries comparisons
i goes on the loop from 1 ... RETRIES, so the right comparison to do is i < RETRIES.

The third change, the i == 0 can never happen, i starts at 1. Anyway the right comparison is i < RETRIES, because i == 1 will cause a continue even when RETRIES is 1, and that will exit the loop without a proper response/content variables, raising an error on the return.